Rename LoginNoSitesError to LoginNoSites to reflect context properly
