+++ WordPress/src/main/java/org/wordpress/android/modules/AppComponent.java
@@ -25,7 +25,7 @@ import org.wordpress.android.ui.accounts.LoginMagicLinkInterceptActivity;
 import org.wordpress.android.ui.accounts.PostSignupInterstitialActivity;
 import org.wordpress.android.ui.accounts.SignupEpilogueActivity;
 import org.wordpress.android.ui.accounts.login.LoginEpilogueFragment;
-import org.wordpress.android.ui.accounts.login.LoginNoSitesErrorFragment;
+import org.wordpress.android.ui.accounts.login.LoginNoSitesFragment;
 import org.wordpress.android.ui.accounts.login.LoginPrologueFragment;
 import org.wordpress.android.ui.accounts.login.LoginSiteCheckErrorFragment;
 import org.wordpress.android.ui.accounts.signup.SignupEpilogueFragment;
@@ -668,7 +668,7 @@ public interface AppComponent extends AndroidInjector<WordPress> {
 
     void inject(LoginSiteCheckErrorFragment object);
 
-    void inject(LoginNoSitesErrorFragment object);
+    void inject(LoginNoSitesFragment object);
 
     void inject(UserProfileBottomSheetFragment object);
