+++ WordPress/src/main/java/org/wordpress/android/modules/ViewModelModule.java
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModelProvider;
 import org.wordpress.android.ui.deeplinks.DeepLinkingIntentReceiverViewModel;
 import org.wordpress.android.ui.JetpackRemoteInstallViewModel;
 import org.wordpress.android.ui.accounts.LoginViewModel;
-import org.wordpress.android.ui.accounts.login.LoginNoSitesErrorViewModel;
+import org.wordpress.android.ui.accounts.login.LoginNoSitesViewModel;
 import org.wordpress.android.ui.activitylog.list.filter.ActivityLogTypeFilterViewModel;
 import org.wordpress.android.ui.domains.DomainRegistrationMainViewModel;
 import org.wordpress.android.ui.engagement.EngagedPeopleListViewModel;
@@ -506,8 +506,8 @@ abstract class ViewModelModule {
 
     @Binds
     @IntoMap
-    @ViewModelKey(LoginNoSitesErrorViewModel.class)
-    abstract ViewModel loginNoSitesErrorViewModel(LoginNoSitesErrorViewModel viewModel);
+    @ViewModelKey(LoginNoSitesViewModel.class)
+    abstract ViewModel loginNoSitesErrorViewModel(LoginNoSitesViewModel viewModel);
 
     @Binds
     @IntoMap
